backend/app/api/endpoints/auth.py
"""
Authentication Endpoints - GitHub OAuth
"""

import secrets
from datetime import datetime, timedelta
from typing import Optional

from fastapi import APIRouter, HTTPException, Depends, Response, Header
from fastapi.responses import RedirectResponse
from typing import Optional as Opt
import httpx
from jose import jwt

from app.config import get_settings
from app.models.schemas import User, UserResponse, APIResponse
from app.services.persistence import save_users, load_users, save_subscription, load_subscription

router = APIRouter()
settings = get_settings()

# Load users from persistence on startup
users_db: dict[str, User] = load_users()
sessions_db: dict[str, str] = {}  # session_token -> user_id


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    """Create JWT access token with user info embedded"""
    to_encode = data.copy()
    # Default to 30 days for better session persistence
    expire = datetime.utcnow() + (expires_delta or timedelta(days=30))
    to_encode.update({"exp": expire, "iat": datetime.utcnow()})
    return jwt.encode(to_encode, settings.secret_key, algorithm="HS256")


def decode_access_token(token: str) -> Optional[dict]:
    """Decode JWT access token"""
    try:
        return jwt.decode(token, settings.secret_key, algorithms=["HS256"])
    except Exception:
        return None


async def get_current_user(authorization: str = None) -> Optional[User]:
    """Get current authenticated user from JWT token"""
    if not authorization:
        return None
    
    # Handle both "Bearer token" and just "token" formats
    token = authorization.replace("Bearer ", "") if authorization.startswith("Bearer ") else authorization
    
    payload = decode_access_token(token)
    
    if not payload or "user_id" not in payload:
        return None
    
    user_id = payload["user_id"]
    
    # First try to get from in-memory store
    user = users_db.get(user_id)
    
    # If not in memory but token is valid, reconstruct user from token payload
    if not user and payload.get("user_data"):
        user_data = payload["user_data"]
        user = User(
            id=user_id,
            github_id=user_data.get("github_id"),
            username=user_data.get("username"),
            email=user_data.get("email"),
            avatar_url=user_data.get("avatar_url"),
            access_token=user_data.get("access_token", ""),
        )
        # Store back in memory for future requests
        users_db[user_id] = user
        # Save to persistence
        save_users(users_db)
    
    return user


@router.get("/github")
async def github_auth():
    """Initiate GitHub OAuth flow"""
    state = secrets.token_urlsafe(32)
    
    github_auth_url = (
        f"https://github.com/login/oauth/authorize"
        f"?client_id={settings.github_client_id}"
        f"&redirect_uri={settings.github_redirect_uri}"
        f"&scope=repo,read:user,user:email"
        f"&state={state}"
    )
    
    return {"auth_url": github_auth_url, "state": state}


@router.get("/github/callback")
async def github_callback(code: str, state: str):
    """Handle GitHub OAuth callback"""
    # Exchange code for access token
    async with httpx.AsyncClient() as client:
        token_response = await client.post(
            "https://github.com/login/oauth/access_token",
            data={
                "client_id": settings.github_client_id,
                "client_secret": settings.github_client_secret,
                "code": code,
                "redirect_uri": settings.github_redirect_uri,
            },
            headers={"Accept": "application/json"}
        )
        
        if token_response.status_code != 200:
            raise HTTPException(status_code=400, detail="Failed to get access token")
        
        token_data = token_response.json()
        access_token = token_data.get("access_token")
        
        if not access_token:
            raise HTTPException(status_code=400, detail="No access token received")
        
        # Get user info from GitHub
        user_response = await client.get(
            "https://api.github.com/user",
            headers={
                "Authorization": f"Bearer {access_token}",
                "Accept": "application/json"
            }
        )
        
        if user_response.status_code != 200:
            raise HTTPException(status_code=400, detail="Failed to get user info")
        
        github_user = user_response.json()
        
        # Get user email
        emails_response = await client.get(
            "https://api.github.com/user/emails",
            headers={
                "Authorization": f"Bearer {access_token}",
                "Accept": "application/json"
            }
        )
        
        primary_email = None
        if emails_response.status_code == 200:
            emails = emails_response.json()
            for email in emails:
                if email.get("primary"):
                    primary_email = email.get("email")
                    break
    
    # Create or update user
    user_id = f"user_{github_user['id']}"
    
    # Check if existing user (preserve subscription tier)
    existing_user = users_db.get(user_id)
    existing_tier = existing_user.subscription_tier if existing_user else "free"
    existing_rzp_id = existing_user.razorpay_customer_id if existing_user else None
    
    user = User(
        id=user_id,
        github_id=github_user["id"],
        username=github_user["login"],
        email=primary_email,
        avatar_url=github_user.get("avatar_url"),
        access_token=access_token,
        subscription_tier=existing_tier,
        razorpay_customer_id=existing_rzp_id,
    )
    
    users_db[user_id] = user
    
    # Save to persistence
    save_users(users_db)
    
    # Initialize free subscription if new user
    if not existing_user:
        from datetime import timezone as tz
        sub_data = {
            "user_id": user_id,
            "tier": "free",
            "currency": "INR",
            "status": "active",
            "usage": {"walkthroughs": 0, "signals": 0, "provenance": 0, "explains": 0, "repos": 0},
            "created_at": datetime.now(tz.utc).isoformat(),
            "updated_at": datetime.now(tz.utc).isoformat(),
        }
        save_subscription(sub_data)
    
    # Create session token with user data embedded
    session_token = create_access_token({
        "user_id": user_id,
        "user_data": {
            "github_id": github_user["id"],
            "username": github_user["login"],
            "email": primary_email,
            "avatar_url": github_user.get("avatar_url"),
            "access_token": access_token,
        }
    })
    
    # Redirect to frontend with token (302 is standard for OAuth)
    return RedirectResponse(
        url=f"{settings.frontend_url}/auth/callback?token={session_token}",
        status_code=302
    )


@router.get("/me")
async def get_me(authorization: Opt[str] = Header(None, alias="Authorization")):
    """Get current user info"""
    user = await get_current_user(authorization)
    
    if not user:
        raise HTTPException(status_code=401, detail="Not authenticated")
    
    # Get subscription info
    sub = load_subscription(user.id)
    tier = sub.get("tier", "free") if sub else user.subscription_tier
    
    return UserResponse(
        id=user.id,
        username=user.username,
        email=user.email,
        avatar_url=user.avatar_url,
        subscription_tier=tier,
    )


@router.post("/logout")
async def logout(authorization: Opt[str] = Header(None, alias="Authorization")):
    """Logout current user"""
    # In production, invalidate the token
    return APIResponse(success=True, message="Logged out successfully")


@router.post("/refresh")
async def refresh_token(authorization: Opt[str] = Header(None, alias="Authorization")):
    """Refresh access token"""
    user = await get_current_user(authorization)
    
    if not user:
        raise HTTPException(status_code=401, detail="Not authenticated")
    
    # Create new token with extended expiry and user data
    new_token = create_access_token({
        "user_id": user.id,
        "user_data": {
            "github_id": user.github_id,
            "username": user.username,
            "email": user.email,
            "avatar_url": user.avatar_url,
            "access_token": user.access_token,
        }
    })
    
    # Get subscription tier
    sub = load_subscription(user.id)
    tier = sub.get("tier", "free") if sub else user.subscription_tier
    
    return {
        "token": new_token,
        "user": UserResponse(
            id=user.id,
            username=user.username,
            email=user.email,
            avatar_url=user.avatar_url,
            subscription_tier=tier,
        )
    }


@router.get("/verify")
async def verify_token(authorization: Opt[str] = Header(None, alias="Authorization")):
    """Verify if token is valid"""
    user = await get_current_user(authorization)
    
    if not user:
        raise HTTPException(status_code=401, detail="Invalid or expired token")
    
    # Get subscription tier
    sub = load_subscription(user.id)
    tier = sub.get("tier", "free") if sub else user.subscription_tier
    
    return {
        "valid": True,
        "user": UserResponse(
            id=user.id,
            username=user.username,
            email=user.email,
            avatar_url=user.avatar_url,
            subscription_tier=tier,
        )
    }

