backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
asyncpg==0.30.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
billiard==4.2.4
boto3>=1.34.0
botocore>=1.34.0
build==1.4.0
celery==5.3.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comtypes==1.4.15; sys_platform == 'win32'
cryptography==46.0.4
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
durationpy==0.10
ecdsa==0.19.1
edge-tts>=7.0.0
fastapi==0.109.0
filelock==3.20.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0

googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.0
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
httpx-sse==0.4.3
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.6.2
kubernetes==35.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.2.1
numpy>=1.22.4,<2
oauthlib==3.3.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==24.2
passlib==1.7.4
posthog>=3.0.0,<4.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
pulsar-client==3.10.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyGithub==2.1.1
Pygments==2.19.2
PyJWT==2.11.0
PyNaCl==1.6.2
pypiwin32==223; sys_platform == 'win32'
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pyttsx3==2.90; sys_platform == 'win32'
pywin32==311; sys_platform == 'win32'
PyYAML==6.0.3
redis==5.0.1
razorpay>=1.4.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.2
rsa==4.9.1
shellingham==1.5.4
six==1.17.0

sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
tqdm==4.67.3
tree-sitter==0.24.0
tree-sitter-go==0.23.4
tree-sitter-java==0.23.5
tree-sitter-javascript==0.23.1
tree-sitter-python==0.23.6
tree-sitter-rust==0.23.2
tree-sitter-typescript==0.23.2
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.27.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
